requirements.txt.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu  # Use faiss-cpu for Railway (it doesn't have GPUs)
groq
requests